backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
python-multipart
httpx
python-dotenv
Flask==3.0.0
flask-cors
psycopg2-binary
boto3
